backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.2.3
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.0.50
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.7.0
matplotlib>=3.9.0
seaborn>=0.13.2
numpy>=1.26.4
openpyxl>=3.1.2